--- src/pages/Cart.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const Cart = () => {
-  return (
-    <div>Cart</div>
-  )
-}
