--- src/pages/Login.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const Login = () => {
-  return (
-    <div>Login</div>
-  )
-}
