--- src/pages/Products.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const Products = () => {
-  return (
-    <div>Products</div>
-  )
-}
