--- src/styles/globalStyles.js
@@ -1,12 +1,0 @@
-import { createGlobalStyle } from 'styled-components';
- 
-const GlobalStyle = createGlobalStyle`
-  body {
-    margin: 0;
-    padding: 0;
-    background: white;
-    font-family: Open-Sans, Helvetica, Sans-Serif;
-  }
-`
- 
-export default GlobalStyle;
