--- src/styles/pages/Home.styles.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const Main = styled.div`
-  display: flex;
-  height: 500px;
-  flex-direction: column;
-  align-items: center;
-  justify-content:center
-`;
